userscript update

+++ initramfs/sbin/userscript.sh
@@ -4,23 +4,6 @@ exec > /data/local/userscript.log 2>&1
 /sbin/busybox mount -o remount,rw / /
 /sbin/busybox mount -o remount,rw /dev/block/mmcblk0p9 /system
 FILESIZE=$(/sbin/busybox cat /data/local/LoStKernel-Ver|wc -c)
-if /sbin/busybox test "$FILESIZE" -ge "10000"; then
-	/sbin/busybox rm /data/local/LoStKernel-Ver
-fi
-
-/sbin/busybox echo "Starting Boot Script - " $(/sbin/busybox date) >> /data/local/LoStKernel-Ver
-/sbin/busybox echo $(/sbin/busybox uname -a) >> /data/local/LoStKernel-Ver
-
-if /sbin/busybox test ! -e "/system/xbin/busybox"; then
-installbb
-fi
-
-if /sbin/busybox test ! -e "/system/bin/su"; then
-	if /sbin/busybox test ! -e "/system/xbin/su"; then
-	installsu
-	fi
-fi
-
 installbb() 
 {
 /sbin/busybox --install -s /system/xbin
@@ -40,6 +23,23 @@ installsu()
 /sbin/busybox rm -rf /sbin/su
 /sbin/busybox echo "Installed SU Binary" >> /data/local/LoStKernel-Ver
 }
+
+if /sbin/busybox test "$FILESIZE" -ge "10000"; then
+	/sbin/busybox rm /data/local/LoStKernel-Ver
+fi
+
+/sbin/busybox echo "Starting Boot Script - " $(/sbin/busybox date) >> /data/local/LoStKernel-Ver
+/sbin/busybox echo $(/sbin/busybox uname -a) >> /data/local/LoStKernel-Ver
+
+if /sbin/busybox test ! -e "/system/xbin/busybox"; then
+installbb
+fi
+
+if /sbin/busybox test ! -e "/system/bin/su"; then
+	if /sbin/busybox test ! -e "/system/xbin/su"; then
+	installsu
+	fi
+fi
 
 if /sbin/busybox test ! -f "/system/app/Superuser.apk"; then
 	if /sbin/busybox test ! -f "/data/app/Superuser.apk"; then
@@ -86,13 +86,15 @@ fi
 # remove bad BB installs
 /sbin/busybox ls -1 /system/bin | while read line
   do
-    if /sbin/busybox test "$(/sbin/busybox basename `/sbin/busybox readlink /system/bin/$line`)" = "busybox"; then
-		/sbin/busybox rm /system/bin/${line}
-		/sbin/busybox echo "Removing Bad BusyBox Install - /system/bin/"${line} >> /data/local/LoStKernel-Ver
-	fi
-    if /sbin/busybox test "$(/sbin/busybox basename `/sbin/busybox readlink /system/bin/$line`)" = "recovery"; then
-		/sbin/busybox rm /system/bin/${line}
-		/sbin/busybox echo "Removing Bad BusyBox Install - /system/bin/"${line} >> /data/local/LoStKernel-Ver
+    if /sbin/busybox test -L "/system/bin/$line"; then
+        if /sbin/busybox test "$(/sbin/busybox basename `/sbin/busybox readlink /system/bin/$line`)" = "busybox"; then
+			/sbin/busybox rm /system/bin/${line}
+			/sbin/busybox echo "Removing Bad BusyBox Install - /system/bin/"${line} >> /data/local/LoStKernel-Ver
+		fi
+		if /sbin/busybox test "$(/sbin/busybox basename `/sbin/busybox readlink /system/bin/$line`)" = "recovery"; then
+			/sbin/busybox rm /system/bin/${line}
+			/sbin/busybox echo "Removing Bad BusyBox Install - /system/bin/"${line} >> /data/local/LoStKernel-Ver
+		fi
 	fi
   done
 # Relink Toolbox
